+++ PyeonHae/PyeonHae/Views/Review/ReviewHome.swift
@@ -6,7 +6,6 @@
 //
 
 import SwiftUI
-import WrappingHStack
 
 // TODO: 리뷰셀 넓이
 
@@ -59,7 +58,11 @@ struct ReviewHome: View {
         ScrollView {
             VStack(spacing: 0) {
                 Spacer().frame(height: 10)
-                filterButton()
+                FilterView(
+                    filterDatas: filterDatas,
+                    showFilter: $showFilter,
+                    selectedElements: $selectedElements
+                )
                 HStack {
                     Spacer().frame(width: 20)
                     Text("새로운 리뷰 14개")
@@ -94,104 +97,5 @@ struct ReviewHome: View {
             Text("팔로우 탭")
             Spacer()
         }
-    }
-    
-    @ViewBuilder
-    private func filterButton() -> some View {
-        HStack(spacing: 4) {
-            Spacer().frame(width: 20)
-            Text("필터")
-                .font(.pretendard(.bold, 14))
-                .foregroundColor(.grayscale100)
-            Image(name: showFilter ? .arrowUp : .arrowDown)
-            Spacer()
-            ForEach(Array(selectedElements.enumerated()), id: \.element) { index, element in
-                if index < 3 {
-                    if index != 0 {
-                        Image(name: .grayCircle)
-                            .padding(.horizontal, 8)
-                    }
-                    Text(element)
-                        .font(.pretendard(.bold, 12))
-                        .foregroundColor(.grayscale85)
-                }
-            }
-            if selectedElements.count > 3 {
-                Text("외\(selectedElements.count - 3)")
-                    .font(.pretendard(.regular, 12))
-                    .foregroundColor(.grayscale85)
-                    .padding(.leading, 8)
-            }
-            Spacer().frame(width: 20)
-        }
-        .frame(height: 32)
-        .onTapGesture {
-            showFilter.toggle()
-        }
-        if showFilter {
-            filterView()
-                .background(Color.grayscale10)
-        }
-    }
-    
-    private func filterView() -> some View {
-        VStack(spacing: 0) {
-            ForEach(Array(filterDatas.enumerated()), id: \.element) { index, data in
-                VStack(alignment: .leading) {
-                    Text(data.category)
-                        .font(.pretendard(.bold, 12))
-                        .foregroundColor(.grayscale85)
-                    WrappingHStack(data.elements, id: \.self, lineSpacing: 6) { element in
-                        customButton(element)
-                    }
-                }
-                .padding(
-                    EdgeInsets(
-                        top: index == 0 ? 16 : 20,
-                        leading: 20,
-                        bottom: index == filterDatas.count - 1 ? 16 : 0,
-                        trailing: 20
-                    )
-                )
-            }
-            .frame(width: UIWindow().screen.bounds.width, alignment: .leading)
-        }
-    }
-    
-    private func customButton(_ text: String) -> some View {
-        Text(text)
-            .font(.pretendard(isSelectedElement(text) ? .semiBold : .regular, 14))
-            .foregroundColor(isSelectedElement(text) ? Color.red100 : Color.grayscale100)
-            .padding(EdgeInsets(top: 7.5, leading: 10, bottom: 7.5, trailing: 10))
-            .cornerRadius(100)
-            .background(
-                RoundedRectangle(cornerRadius: 100)
-                    .fill(isSelectedElement(text) ? Color.red100.opacity(0.1) : Color.white)
-            )
-            .overlay(
-                RoundedRectangle(cornerRadius: 100)
-                    .stroke(isSelectedElement(text) ? Color.red100 : Color.grayscale30, lineWidth: 1)
-            )
-            .onTapGesture {
-                if selectedElements.contains(text) {
-                    selectedElements.removeAll(where: { $0 == text })
-                } else {
-                    selectedElements.append(text)
-                }
-            }
-    }
-    
-    private func isSelectedElement(_ element: String) -> Bool {
-        return selectedElements.contains(element)
-    }
-}
-
-struct FilterData: Hashable {
-    let category: String
-    let elements: [String]
-    
-    init(category: String, elements: [String]) {
-        self.category = category
-        self.elements = elements
     }
 }
